[FIX] setup: support PEP 660 and setuptools >= 64
Allow building a wheel or an sdist through the pep517
interface, taking care of the required pre-processing
to copy addons in odoo/addons.

Force the setuptools compat mode for editable installs.
This mode is the simplest, since it just installs a .pth
files which extends PYTHONPATH with the Odoo root directory.